Move converters into JSONConverters namespace
