+++ S2VX.Game/Story/JSONConverters/Vector2Converter.cs
@@ -3,7 +3,7 @@ using Newtonsoft.Json.Linq;
 using osuTK;
 using System;
 
-namespace S2VX.Game.Story {
+namespace S2VX.Game.Story.JSONConverters {
     // Special converter for serialization since we only care about the X and Y values of Vector2
     public class Vector2Converter : JsonConverter<Vector2> {
         public override void WriteJson(JsonWriter writer, Vector2 vector2, JsonSerializer serializer) {
